Behold I have a sprite following the mouse.
Found out that there are several not good places to do this building ghost sprite following the mouse. In the button code was bad cause zero zero was the left corner of the container. Cursor manager probably would have worked but I was getting huge numbers for position when i would edge scroll. Not sure what was happening there. Seems like an issue with moving the camera's position then doing a bunch of additional math to find a target pos for the ghost building. Now we just simply use the position of camera + local mouse position.