back to using navigation mesh agents
Seems like combining the formation manager's logic of guaranteed tile for every unit in the selection is pretty much how aoe2 does it. And it works well to keep the units from clumping and never reaching their goal. I think Starcraft 2 does a lot of pushing for friendlies once they are stopped. Unsure how to replicate that in a good way with avoidance. Right now it is hit or miss wether the stopped unit will respond at all to the pushing of the unit still trying to get to the goal