Navigation using obstructions and the NavServer
Okay I think this is good enough to merge into main now. I wasn't sure if this was going to work but I have cell blocking working all off of procedural nav mesh objects. So units will add their obstruction outline to a dictionary and update its polygon as they complete navigation. So they block a cell when they are not moving but unblock that cell once they start moving to a new cell which they will block once they arrive. This is paired with the formation_handler's guarantee that each unit will  have their own cell to path to.