Cell Block module so anything can block a cell
Cell blocking is now in its own script so I don't have to inherit from some object for everything that needs to block a cell. Buildings and units don't share an easily editable ancestor so this works better. Also so reorganizing of functions and variables into regions and alphabetical order